Pulser/Delays use bindings.Comment out some network commands
For the sake of reducing the amount of code I used
cocoa bindings to connect the pulser/delays text boxes
and their corresponding model values. I also had all of the
fire/stop/load buttons go to one of a few functions which then
multiplexes to the correct model command. So instead of there being
a function fireTellie, and fireSmellie and firePulser. THere's now
just firePulser which handles all 3

ALSO. IMPORTANT. I commented out a few network communication calls so
that I could test the buttons without the being connected to the server